+++ frontend/src/pages/blog/Blog.jsx
@@ -12,6 +12,7 @@ import {
 import { motion } from 'framer-motion';
 import BackToTopButton from '@/components/utils/BackToTopButton';
 import Spinner from '@/components/spinner/Spinner';
+import Pagination from '@/components/utils/Pagination';
 
 export default function Blog() {
   const [posts, setPosts] = useState([]);
@@ -61,30 +62,6 @@ export default function Blog() {
       month: 'long',
       day: 'numeric',
     });
-  };
-
-  // Manage page changes
-  const handlePageChange = (pageNum) => {
-    if (pageNum >= 1 && pageNum <= totalPages && pageNum !== currentPage) {
-      setCurrentPage(pageNum);
-    }
-  };
-
-  // Generate visible page numbers
-  const generatePageNumbers = () => {
-    const maxVisible = 5;
-    let start = Math.max(1, currentPage - Math.floor(maxVisible / 2));
-    let end = Math.min(totalPages, start + maxVisible - 1);
-
-    if (end - start < maxVisible - 1) {
-      start = Math.max(1, end - maxVisible + 1);
-    }
-
-    const pages = [];
-    for (let i = start; i <= end; i++) {
-      pages.push(i);
-    }
-    return pages;
   };
 
   // Loading (show spinner first)
@@ -278,58 +255,13 @@ export default function Blog() {
           </div>
 
           {/* Pagination */}
-          {totalPages > 1 && (
-            <div className='flex flex-col items-center space-y-4'>
-              <div className='flex items-center space-x-2'>
-                {/* Prev */}
-                <button
-                  onClick={() => handlePageChange(currentPage - 1)}
-                  disabled={currentPage === 1}
-                  className={`px-4 py-2 rounded-lg transition-all ${
-                    currentPage === 1
-                      ? 'opacity-50 cursor-not-allowed bg-gray-700'
-                      : 'bg-purple-600 hover:bg-purple-500'
-                  }`}
-                >
-                  <FaChevronLeft />
-                </button>
-
-                {/* Numbers */}
-                {generatePageNumbers().map((pageNum) => (
-                  <button
-                    key={pageNum}
-                    onClick={() => handlePageChange(pageNum)}
-                    className={`w-10 h-10 rounded-lg transition-all duration-200 ${
-                      currentPage === pageNum
-                        ? 'bg-purple-600 text-white scale-110'
-                        : 'bg-white/[0.08] text-white/70 hover:bg-purple-600'
-                    }`}
-                  >
-                    {pageNum}
-                  </button>
-                ))}
-
-                {/* Next */}
-                <button
-                  onClick={() => handlePageChange(currentPage + 1)}
-                  disabled={currentPage === totalPages}
-                  className={`px-4 py-2 rounded-lg transition-all ${
-                    currentPage === totalPages
-                      ? 'opacity-50 cursor-not-allowed bg-gray-700'
-                      : 'bg-purple-600 hover:bg-purple-500'
-                  }`}
-                >
-                  <FaChevronRight />
-                </button>
-              </div>
-
-              <div className='text-sm text-gray-400'>
-                Showing {(currentPage - 1) * postsPerPage + 1} to{' '}
-                {Math.min(currentPage * postsPerPage, totalPosts)} of{' '}
-                {totalPosts} posts
-              </div>
-            </div>
-          )}
+          <Pagination 
+            currentPage={currentPage} 
+            totalPages={totalPages} 
+            totalItems={totalPosts} 
+            itemsPerPage={postsPerPage} 
+            setCurrentPage={setCurrentPage}
+          />
         </div>
       </motion.section>
       {/* Floating Back-to-Top Button Component */}
